# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ShopPro E-Commerce API",
    "version": "1.0.0",
    "description": "OpenAPI documentation for the ShopPro backend. All /v1/api endpoints require x-api-key. Protected user/admin endpoints also require Authorization and x-client-id."
  },
  "servers": [
    {
      "url": "http://localhost:5500",
      "description": "Local backend"
    }
  ],
...
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/v1/api/shop/handlerRefreshToken": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh access token from httpOnly refresh cookie",
        "operationId": "refreshToken",
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthSuccess"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Required for all /v1/api endpoints."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token returned by login endpoints."
      },
      "ClientIdAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-client-id",
        "description": "Current user/shop id."
      },
      "AuthKindAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-auth-kind",
        "description": "Use user for customer APIs and admin for admin/shop APIs."
      }
    },
...
    "responses": {
...
      "AuthSuccess": {
        "description": "Authentication success response",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "metadata": {
                      "$ref": "#/components/schemas/AuthResponse"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "Error": {
        "description": "Standard error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "example": 200
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "code": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "additionalProperties": true
          },
          "shop": {
            "type": "object",
            "additionalProperties": true
          },
          "tokens": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}